Student logbook
The project is built using ASP.NET Core MVC technology to implement server software with web interface and machine learning technologies. 
This project simulates the work of the application of student journal with grades, lists of students who are on the verge of expulsion. With the help of machine learning technology is the calculation of the preliminary assessment of the student for the exam.